// apps/frontend2/src/api/models/jobListingDTO.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { JobListingDTOSource } from './jobListingDTOSource';
import type { JobListingDTODescription } from './jobListingDTODescription';
import type { JobListingDTOPostedAt } from './jobListingDTOPostedAt';

export interface JobListingDTO {
  id: number;
  title: string;
  application_url: string;
  company: string;
  source: JobListingDTOSource;
  description: JobListingDTODescription;
  posted_at: JobListingDTOPostedAt;
  scraped_at: string;
  updated_at: string;
}